// src/types.ts
export interface EmpresaConfig {
  nombreEmpresa: string;
  rif: string;
  direccionFiscal: string;
  telefono: string;
  logoUrl?: string;
  tasaCambio: number; // Bolívares por USD (ej. 36.50)
  fechaTasa: string;
  ultimaActualizacionTasa?: string;
  nombreTienda1: string;
  nombreTienda2: string;
  nombreOficina: string;
}